refactoring channel list view to use fragments for detail view
The existing UI architecture for main view and list view
remains the same, but instead of attaching detail layout directly
(as in previous implementation) we create a fragment container
view and attach the fragment in its dedicated detail there.

This allows for:

- reusing existing gesture recognition and animations, with no
  changes

- more flexibility for future tablet layouts, as we now use
  fragments for detail views

- new detail views can be implemented within fragments and use
  MVVM pattern

This commit also adds a LegacyDatailFragment, which is a wrapper
around existing detail views, so that they can be seamlessly
integrated into the new architecture.